Startup hook to initialize FBServices
