+++ docroot/WEB-INF/src/portal.properties
@@ -1,1 +1,2 @@
+application.startup.events=net.indaba.lostandfound.hook.FirebaseServicesInitAction
 value.object.listener.com.liferay.portal.kernel.model.Group=net.indaba.lostandfound.hook.LFGroupModelListener
